♻️refactor: create day off schedule adjustment controller
